+++ network-store-server/src/main/java/com/powsybl/network/store/server/NetworkStoreRepository.java
@@ -24,7 +24,6 @@ import com.powsybl.network.store.server.exceptions.JsonApiErrorResponseException
 import com.powsybl.network.store.server.exceptions.UncheckedSqlException;
 import com.powsybl.network.store.server.json.PermanentLimitSqlData;
 import com.powsybl.network.store.server.json.TemporaryLimitSqlData;
-import com.powsybl.network.store.server.migration.V211LimitsQueryCatalog;
 import com.powsybl.ws.commons.LogUtils;
 import lombok.Getter;
 import org.apache.commons.lang3.mutable.MutableInt;
@@ -44,6 +43,7 @@ import static com.powsybl.network.store.server.Mappings.*;
 import static com.powsybl.network.store.server.QueryCatalog.*;
 import static com.powsybl.network.store.server.Utils.bindAttributes;
 import static com.powsybl.network.store.server.Utils.bindValues;
+import static com.powsybl.network.store.server.migration.V211LimitsMigration.*;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -193,31 +193,25 @@ public class NetworkStoreRepository {
         }
     }
 
-    private static void executeWithoutAutoCommit(Connection connection, SqlExecutor executor, Boolean withException) throws SQLException {
+    private static void executeWithoutAutoCommit(Connection connection, SqlExecutor executor) throws SQLException {
         connection.setAutoCommit(false);
         try {
             executor.execute(connection);
             connection.commit();
         } catch (Exception e) {
             rollbackQuietly(connection);
-            if (Boolean.TRUE.equals(withException)) {
-                throw new RuntimeException(e);
-            }
+            throw new RuntimeException(e);
         } finally {
             restoreAutoCommitQuietly(connection);
         }
     }
 
-    private void executeWithoutAutoCommit(SqlExecutor executor, Boolean withException) {
+    private void executeWithoutAutoCommit(SqlExecutor executor) {
         try (var connection = dataSource.getConnection()) {
-            executeWithoutAutoCommit(connection, executor, withException);
+            executeWithoutAutoCommit(connection, executor);
         } catch (SQLException e) {
             throw new UncheckedSqlException(e);
         }
-    }
-
-    private void executeWithoutAutoCommit(SqlExecutor executor) {
-        executeWithoutAutoCommit(executor, true);
     }
 
     public void createNetworks(List<Resource<NetworkAttributes>> resources) {
@@ -1799,7 +1793,14 @@ public class NetworkStoreRepository {
 
     public Map<OwnerInfo, LimitsInfos> getLimitsInfos(UUID networkUuid, int variantNum, String columnNameForWhereClause, String valueForWhereClause) {
         //To be removed when limits are fully migrated — should be after v2.13 deployment
-        migrateV211Limits(networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
+        Map<OwnerInfo, List<TemporaryLimitAttributes>> oldTemporaryLimits;
+        Map<OwnerInfo, List<PermanentLimitAttributes>> oldPermanentLimits;
+        oldTemporaryLimits = getV211TemporaryLimits(this, networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
+        oldPermanentLimits = getV211PermanentLimits(this, networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
+
+        if (!oldTemporaryLimits.isEmpty() || !oldPermanentLimits.isEmpty()) {
+            return mergeLimitsIntoLimitsInfos(oldTemporaryLimits, oldPermanentLimits);
+        }
 
         Map<OwnerInfo, List<TemporaryLimitAttributes>> temporaryLimits = getTemporaryLimits(networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
         Map<OwnerInfo, List<PermanentLimitAttributes>> permanentLimits = getPermanentLimits(networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
@@ -1808,7 +1809,13 @@ public class NetworkStoreRepository {
 
     public Map<OwnerInfo, LimitsInfos> getLimitsInfosWithInClause(UUID networkUuid, int variantNum, String columnNameForWhereClause, List<String> valuesForInClause) {
         //To be removed when limits are fully migrated — should be after v2.13 deployment
-        migrateV211LimitsWithInClause(networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
+        Map<OwnerInfo, List<TemporaryLimitAttributes>> oldTemporaryLimits;
+        Map<OwnerInfo, List<PermanentLimitAttributes>> oldPermanentLimits;
+        oldTemporaryLimits = getV211TemporaryLimitsWithInClause(this, networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
+        oldPermanentLimits = getV211PermanentLimitsWithInClause(this, networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
+        if (!oldTemporaryLimits.isEmpty() || !oldPermanentLimits.isEmpty()) {
+            return mergeLimitsIntoLimitsInfos(oldTemporaryLimits, oldPermanentLimits);
+        }
 
         Map<OwnerInfo, List<TemporaryLimitAttributes>> temporaryLimits = getTemporaryLimitsWithInClause(networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
         Map<OwnerInfo, List<PermanentLimitAttributes>> permanentLimits = getPermanentLimitsWithInClause(networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
@@ -1831,7 +1838,6 @@ public class NetworkStoreRepository {
 
     public Map<OwnerInfo, List<TemporaryLimitAttributes>> getTemporaryLimits(UUID networkUuid, int variantNum, String columnNameForWhereClause, String valueForWhereClause) {
         try (var connection = dataSource.getConnection()) {
-            // get from new table
             var preparedStmt = connection.prepareStatement(QueryCatalog.buildTemporaryLimitQuery(columnNameForWhereClause));
             preparedStmt.setString(1, networkUuid.toString());
             preparedStmt.setInt(2, variantNum);
@@ -1927,7 +1933,7 @@ public class NetworkStoreRepository {
     }
 
     public void insertTemporaryLimitsAttributes(Map<OwnerInfo, List<TemporaryLimitAttributes>> temporaryLimits) {
-        executeWithoutAutoCommit(connection -> {
+        try (var connection = dataSource.getConnection()) {
             try (var preparedStmt = connection.prepareStatement(QueryCatalog.buildInsertTemporaryLimitsQuery())) {
                 List<Object> values = new ArrayList<>(5);
                 List<Map.Entry<OwnerInfo, List<TemporaryLimitAttributes>>> list = new ArrayList<>(temporaryLimits.entrySet());
@@ -1949,7 +1955,9 @@ public class NetworkStoreRepository {
                     preparedStmt.executeBatch();
                 }
             }
-        }, false);
+        } catch (SQLException e) {
+            throw new UncheckedSqlException(e);
+        }
     }
 
     public void insertPermanentLimits(Map<OwnerInfo, LimitsInfos> limitsInfos) {
@@ -1959,7 +1967,7 @@ public class NetworkStoreRepository {
     }
 
     public void insertPermanentLimitsAttributes(Map<OwnerInfo, List<PermanentLimitAttributes>> permanentLimits) {
-        executeWithoutAutoCommit(connection -> {
+        try (var connection = dataSource.getConnection()) {
             try (var preparedStmt = connection.prepareStatement(QueryCatalog.buildInsertPermanentLimitsQuery())) {
                 List<Object> values = new ArrayList<>(8);
                 List<Map.Entry<OwnerInfo, List<PermanentLimitAttributes>>> list = new ArrayList<>(permanentLimits.entrySet());
@@ -1978,7 +1986,9 @@ public class NetworkStoreRepository {
                     preparedStmt.executeBatch();
                 }
             }
-        }, false);
+        } catch (SQLException e) {
+            throw new UncheckedSqlException(e);
+        }
     }
 
     protected <T extends LimitHolder & IdentifiableAttributes> void insertLimitsInEquipments(UUID networkUuid, List<Resource<T>> equipments, Map<OwnerInfo, LimitsInfos> limitsInfos) {
@@ -2036,10 +2046,10 @@ public class NetworkStoreRepository {
     }
 
     private void deleteTemporaryLimits(UUID networkUuid, int variantNum, List<String> equipmentIds) {
-        //To be removed when limits are fully migrated — should be after v2.13 deployment
-        deleteV211TemporaryLimits(networkUuid, variantNum, equipmentIds);
-
         try (var connection = dataSource.getConnection()) {
+            //To be removed when limits are fully migrated — should be after v2.13 deployment
+            deleteV211TemporaryLimits(connection, networkUuid, variantNum, equipmentIds);
+
             try (var preparedStmt = connection.prepareStatement(QueryCatalog.buildDeleteTemporaryLimitsVariantEquipmentINQuery(equipmentIds.size()))) {
                 preparedStmt.setString(1, networkUuid.toString());
                 preparedStmt.setInt(2, variantNum);
@@ -2054,10 +2064,10 @@ public class NetworkStoreRepository {
     }
 
     private void deletePermanentLimits(UUID networkUuid, int variantNum, List<String> equipmentIds) {
-        //To be removed when limits are fully migrated — should be after v2.13 deployment
-        deleteV211PermanentLimits(networkUuid, variantNum, equipmentIds);
-
         try (var connection = dataSource.getConnection()) {
+            //To be removed when limits are fully migrated — should be after v2.13 deployment
+            deleteV211PermanentLimits(connection, networkUuid, variantNum, equipmentIds);
+
             try (var preparedStmt = connection.prepareStatement(QueryCatalog.buildDeletePermanentLimitsVariantEquipmentINQuery(equipmentIds.size()))) {
                 preparedStmt.setString(1, networkUuid.toString());
                 preparedStmt.setInt(2, variantNum);
@@ -2935,207 +2945,5 @@ public class NetworkStoreRepository {
 
     public void removeExtensionAttributes(UUID networkId, int variantNum, String identifiableId, String extensionName) {
         extensionHandler.deleteExtensionsFromIdentifiables(networkId, variantNum, Map.of(identifiableId, Set.of(extensionName)));
-    }
-
-    // Methods to migrate V2.11 limits — do not use them in operational code
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public Map<OwnerInfo, List<PermanentLimitAttributes>> getV211PermanentLimitsWithInClause(UUID networkUuid, int variantNum, String columnNameForWhereClause, List<String> valuesForInClause) {
-        if (valuesForInClause.isEmpty()) {
-            return Collections.emptyMap();
-        }
-        try (var connection = dataSource.getConnection()) {
-            var preparedStmt = connection.prepareStatement(V211LimitsQueryCatalog.buildGetV211PermanentLimitWithInClauseQuery(columnNameForWhereClause, valuesForInClause.size()));
-            preparedStmt.setString(1, networkUuid.toString());
-            preparedStmt.setInt(2, variantNum);
-            for (int i = 0; i < valuesForInClause.size(); i++) {
-                preparedStmt.setString(3 + i, valuesForInClause.get(i));
-            }
-
-            return innerGetV211PermanentLimits(preparedStmt);
-        } catch (SQLException e) {
-            throw new UncheckedSqlException(e);
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public Map<OwnerInfo, List<TemporaryLimitAttributes>> getV211TemporaryLimits(UUID networkUuid, int variantNum, String columnNameForWhereClause, String valueForWhereClause) {
-        try (var connection = dataSource.getConnection()) {
-
-            // get from old table
-            var preparedStmt = connection.prepareStatement(V211LimitsQueryCatalog.buildGetV211TemporaryLimitQuery(columnNameForWhereClause));
-            preparedStmt.setString(1, networkUuid.toString());
-            preparedStmt.setInt(2, variantNum);
-            preparedStmt.setString(3, valueForWhereClause);
-
-            return innerGetV211TemporaryLimits(preparedStmt);
-        } catch (SQLException e) {
-            throw new UncheckedSqlException(e);
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public Map<OwnerInfo, List<PermanentLimitAttributes>> getV211PermanentLimits(UUID networkUuid, int variantNum, String columnNameForWhereClause, String valueForWhereClause) {
-        try (var connection = dataSource.getConnection()) {
-            var preparedStmt = connection.prepareStatement(V211LimitsQueryCatalog.buildGetV211PermanentLimitQuery(columnNameForWhereClause));
-            preparedStmt.setString(1, networkUuid.toString());
-            preparedStmt.setInt(2, variantNum);
-            preparedStmt.setString(3, valueForWhereClause);
-
-            return innerGetV211PermanentLimits(preparedStmt);
-        } catch (SQLException e) {
-            throw new UncheckedSqlException(e);
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public Map<OwnerInfo, List<TemporaryLimitAttributes>> getV211TemporaryLimitsWithInClause(UUID networkUuid, int variantNum, String columnNameForWhereClause, List<String> valuesForInClause) {
-        if (valuesForInClause.isEmpty()) {
-            return Collections.emptyMap();
-        }
-        try (var connection = dataSource.getConnection()) {
-            var preparedStmt = connection.prepareStatement(V211LimitsQueryCatalog.buildGetV211TemporaryLimitWithInClauseQuery(columnNameForWhereClause, valuesForInClause.size()));
-            preparedStmt.setString(1, networkUuid.toString());
-            preparedStmt.setInt(2, variantNum);
-            for (int i = 0; i < valuesForInClause.size(); i++) {
-                preparedStmt.setString(3 + i, valuesForInClause.get(i));
-            }
-
-            return innerGetV211TemporaryLimits(preparedStmt);
-        } catch (SQLException e) {
-            throw new UncheckedSqlException(e);
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public void migrateV211Limits(UUID networkId) {
-        Stopwatch stopwatch = Stopwatch.createStarted();
-
-        List<Integer> variantNums = getVariantsInfos(networkId).stream().map(VariantInfos::getNum).toList();
-        variantNums.forEach(variantNum -> {
-            migrateV211Limits(networkId, variantNum, EQUIPMENT_TYPE_COLUMN, ResourceType.LINE.toString());
-            migrateV211Limits(networkId, variantNum, EQUIPMENT_TYPE_COLUMN, ResourceType.TWO_WINDINGS_TRANSFORMER.toString());
-            migrateV211Limits(networkId, variantNum, EQUIPMENT_TYPE_COLUMN, ResourceType.THREE_WINDINGS_TRANSFORMER.toString());
-            migrateV211Limits(networkId, variantNum, EQUIPMENT_TYPE_COLUMN, ResourceType.DANGLING_LINE.toString());
-        });
-
-        stopwatch.stop();
-        LOGGER.info("Limits of network {} migrated in {} ms", networkId, stopwatch.elapsed(TimeUnit.MILLISECONDS));
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public boolean insertNewLimitsAndDeleteV211Ones(UUID networkUuid, int variantNum, Map<OwnerInfo, List<TemporaryLimitAttributes>> oldTemporaryLimits, Map<OwnerInfo, List<PermanentLimitAttributes>> oldPermanentLimits) {
-        if (!oldPermanentLimits.keySet().isEmpty()) {
-            insertPermanentLimitsAttributes(oldPermanentLimits);
-            deleteV211PermanentLimits(networkUuid, variantNum, oldPermanentLimits.keySet().stream().map(OwnerInfo::getEquipmentId).toList());
-        }
-        if (!oldTemporaryLimits.keySet().isEmpty()) {
-            insertTemporaryLimitsAttributes(oldTemporaryLimits);
-            deleteV211TemporaryLimits(networkUuid, variantNum, oldTemporaryLimits.keySet().stream().map(OwnerInfo::getEquipmentId).toList());
-        }
-        return !oldPermanentLimits.keySet().isEmpty() || !oldTemporaryLimits.keySet().isEmpty();
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public void migrateV211Limits(UUID networkUuid, int variantNum, String columnNameForWhereClause, String valueForWhereClause) {
-        Stopwatch stopwatch = Stopwatch.createStarted();
-        Map<OwnerInfo, List<TemporaryLimitAttributes>> oldTemporaryLimits = getV211TemporaryLimits(networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
-        Map<OwnerInfo, List<PermanentLimitAttributes>> oldPermanentLimits = getV211PermanentLimits(networkUuid, variantNum, columnNameForWhereClause, valueForWhereClause);
-        boolean migrated = insertNewLimitsAndDeleteV211Ones(networkUuid, variantNum, oldTemporaryLimits, oldPermanentLimits);
-        stopwatch.stop();
-        if (Boolean.TRUE.equals(migrated)) {
-            LOGGER.info("Limits of {}S of network {}/variantNum={} migrated in {} ms", valueForWhereClause, networkUuid, variantNum, stopwatch.elapsed(TimeUnit.MILLISECONDS));
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    public void migrateV211LimitsWithInClause(UUID networkUuid, int variantNum, String columnNameForWhereClause, List<String> valuesForInClause) {
-        Map<OwnerInfo, List<TemporaryLimitAttributes>> oldTemporaryLimits = getV211TemporaryLimitsWithInClause(networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
-        Map<OwnerInfo, List<PermanentLimitAttributes>> oldPermanentLimits = getV211PermanentLimitsWithInClause(networkUuid, variantNum, columnNameForWhereClause, valuesForInClause);
-        insertNewLimitsAndDeleteV211Ones(networkUuid, variantNum, oldTemporaryLimits, oldPermanentLimits);
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    private Map<OwnerInfo, List<TemporaryLimitAttributes>> innerGetV211TemporaryLimits(PreparedStatement preparedStmt) throws SQLException {
-        try (ResultSet resultSet = preparedStmt.executeQuery()) {
-            Map<OwnerInfo, List<TemporaryLimitAttributes>> map = new HashMap<>();
-            while (resultSet.next()) {
-                OwnerInfo owner = new OwnerInfo();
-                TemporaryLimitAttributes temporaryLimit = new TemporaryLimitAttributes();
-                // In order, from the QueryCatalog.buildTemporaryLimitQuery SQL query :
-                // equipmentId, equipmentType, networkUuid, variantNum, side, limitType, name, value, acceptableDuration, fictitious
-                owner.setEquipmentId(resultSet.getString(1));
-                owner.setEquipmentType(ResourceType.valueOf(resultSet.getString(2)));
-                owner.setNetworkUuid(UUID.fromString(resultSet.getString(3)));
-                owner.setVariantNum(resultSet.getInt(4));
-                temporaryLimit.setOperationalLimitsGroupId(resultSet.getString(5));
-                temporaryLimit.setSide(resultSet.getInt(6));
-                temporaryLimit.setLimitType(LimitType.valueOf(resultSet.getString(7)));
-                temporaryLimit.setName(resultSet.getString(8));
-                temporaryLimit.setValue(resultSet.getDouble(9));
-                temporaryLimit.setAcceptableDuration(resultSet.getInt(10));
-                temporaryLimit.setFictitious(resultSet.getBoolean(11));
-                map.computeIfAbsent(owner, k -> new ArrayList<>());
-                map.get(owner).add(temporaryLimit);
-            }
-            return map;
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    private Map<OwnerInfo, List<PermanentLimitAttributes>> innerGetV211PermanentLimits(PreparedStatement preparedStmt) throws SQLException {
-        try (ResultSet resultSet = preparedStmt.executeQuery()) {
-            Map<OwnerInfo, List<PermanentLimitAttributes>> map = new HashMap<>();
-            while (resultSet.next()) {
-                OwnerInfo owner = new OwnerInfo();
-                PermanentLimitAttributes permanentLimit = new PermanentLimitAttributes();
-                // In order, from the QueryCatalog.buildTemporaryLimitQuery SQL query :
-                // equipmentId, equipmentType, networkUuid, variantNum, side, limitType, name, value, acceptableDuration, fictitious
-                owner.setEquipmentId(resultSet.getString(1));
-                owner.setEquipmentType(ResourceType.valueOf(resultSet.getString(2)));
-                owner.setNetworkUuid(UUID.fromString(resultSet.getString(3)));
-                owner.setVariantNum(resultSet.getInt(4));
-                permanentLimit.setOperationalLimitsGroupId(resultSet.getString(5));
-                permanentLimit.setSide(resultSet.getInt(6));
-                permanentLimit.setLimitType(LimitType.valueOf(resultSet.getString(7)));
-                permanentLimit.setValue(resultSet.getDouble(8));
-
-                map.computeIfAbsent(owner, k -> new ArrayList<>());
-                map.get(owner).add(permanentLimit);
-            }
-            return map;
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    private void deleteV211TemporaryLimits(UUID networkUuid, int variantNum, List<String> equipmentIds) {
-        try (var connection = dataSource.getConnection()) {
-            try (var preparedStmt = connection.prepareStatement(V211LimitsQueryCatalog.buildDeleteV211TemporaryLimitsVariantEquipmentINQuery(equipmentIds.size()))) {
-                preparedStmt.setString(1, networkUuid.toString());
-                preparedStmt.setInt(2, variantNum);
-                for (int i = 0; i < equipmentIds.size(); i++) {
-                    preparedStmt.setString(3 + i, equipmentIds.get(i));
-                }
-                preparedStmt.executeUpdate();
-            }
-        } catch (SQLException e) {
-            throw new UncheckedSqlException(e);
-        }
-    }
-
-    //To be deprecated when limits are fully migrated — should be after v2.13 deployment
-    private void deleteV211PermanentLimits(UUID networkUuid, int variantNum, List<String> equipmentIds) {
-        try (var connection = dataSource.getConnection()) {
-            try (var preparedStmt = connection.prepareStatement(V211LimitsQueryCatalog.buildDeleteV211PermanentLimitsVariantEquipmentINQuery(equipmentIds.size()))) {
-                preparedStmt.setString(1, networkUuid.toString());
-                preparedStmt.setInt(2, variantNum);
-                for (int i = 0; i < equipmentIds.size(); i++) {
-                    preparedStmt.setString(3 + i, equipmentIds.get(i));
-                }
-                preparedStmt.executeUpdate();
-            }
-        } catch (SQLException e) {
-            throw new UncheckedSqlException(e);
-        }
     }
 }
