+++ network-store-server/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -92,6 +92,6 @@ databaseChangeLog:
       file: changesets/changelog_20241121T110000Z.xml
       relativeToChangelogFile: true
 
-#  - include:
-#      file: changesets/changelog_20241223T130000Z.xml
-#      relativeToChangelogFile: true
+  - include:
+      file: changesets/changelog_20241223T130000Z.xml
+      relativeToChangelogFile: true
